Handle edge cases in tool panel UI, move a settings function
