+++ src/js/tools/tools.js
@@ -1,4 +1,4 @@
-import {getSettings} from './settings-ui';
+import {getSettings, handleSettingsHeaderClick} from './settings-ui';
 import version from '../version';
 
 const style = `
@@ -38,15 +38,18 @@ const style = `
       cursor: pointer;
     }
 
-    #tools li:hover, #download li:hover,
-    #tools li.active {
+    #tools li:hover,
+    #tools li.active,
+    #download li:hover {
       background: #DDD;
     }
 
-    #tools li.ideo-disabled:hover,
-    #download li.ideo-disabled:hover,
-    #tools li.active.ideo-disabled {
+    #tools li.ideo-disabled,
+    #tools li.active.ideo-disabled,
+    #download li.ideo-disabled {
       background: #FFF;
+      color: #CCC;
+      cursor: default;
     }
 
     #download {
@@ -103,8 +106,8 @@ const style = `
     #download {
       position: absolute;
       width: 150px;
-      top: -8px;
-      right: 145px;
+      top: -2px;
+      right: 120px;
       z-index: 8000;
       background: white;
       margin: 0;
@@ -207,42 +210,12 @@ const style = `
       border-top: none;
     }
 
-    #gear .ideo-disabled {
-      color: #CCC;
-      cursor: default;
-    }
   </style>`;
 
 // eslint-disable-next-line max-len
 const gearIcon = '<svg viewBox="0 0 512 512"><path fill="#AAA" d="M444.788 291.1l42.616 24.599c4.867 2.809 7.126 8.618 5.459 13.985-11.07 35.642-29.97 67.842-54.689 94.586a12.016 12.016 0 0 1-14.832 2.254l-42.584-24.595a191.577 191.577 0 0 1-60.759 35.13v49.182a12.01 12.01 0 0 1-9.377 11.718c-34.956 7.85-72.499 8.256-109.219.007-5.49-1.233-9.403-6.096-9.403-11.723v-49.184a191.555 191.555 0 0 1-60.759-35.13l-42.584 24.595a12.016 12.016 0 0 1-14.832-2.254c-24.718-26.744-43.619-58.944-54.689-94.586-1.667-5.366.592-11.175 5.459-13.985L67.212 291.1a193.48 193.48 0 0 1 0-70.199l-42.616-24.599c-4.867-2.809-7.126-8.618-5.459-13.985 11.07-35.642 29.97-67.842 54.689-94.586a12.016 12.016 0 0 1 14.832-2.254l42.584 24.595a191.577 191.577 0 0 1 60.759-35.13V25.759a12.01 12.01 0 0 1 9.377-11.718c34.956-7.85 72.499-8.256 109.219-.007 5.49 1.233 9.403 6.096 9.403 11.723v49.184a191.555 191.555 0 0 1 60.759 35.13l42.584-24.595a12.016 12.016 0 0 1 14.832 2.254c24.718 26.744 43.619 58.944 54.689 94.586 1.667 5.366-.592 11.175-5.459 13.985L444.788 220.9a193.485 193.485 0 0 1 0 70.2zM336 256c0-44.112-35.888-80-80-80s-80 35.888-80 80 35.888 80 80 80 80-35.888 80-80z"/></svg>';
 // Font Awesome Free 5.2.0 by @fontawesome - https://fontawesome.com
 // License - https://fontawesome.com/license (Icons: CC BY 4.0, Fonts: SIL OFL 1.1, Code: MIT License)
-
-function handleSettingsHeaderClick(ideo) {
-  var links = document.querySelectorAll('li.ideo-settings-header > a');
-  links.forEach(function(link) {
-    link.addEventListener('click', function(event) {
-      var targetLink = event.target,
-        targetId;
-
-      // Don't scroll
-      event.preventDefault();
-
-      // Deactivate all tabs
-      links.forEach(function(link2) {
-        link2.parentElement.classList.remove('active');
-      });
-      document.querySelectorAll('.tab-content > div').forEach(function(div) {
-        div.classList.remove('active');
-      });
-
-      // Activate selected tab
-      targetLink.parentElement.classList += ' active';
-      targetId = targetLink.href.split('#')[1];
-      document.getElementById(targetId).classList += ' active';
-    });
-  });
-}
 
 function deactivate(items) {
   items.forEach(item => {item.classList.remove('active');});
@@ -264,26 +237,22 @@ function closeTools() {
 function handleHideForHoverables(trigger, tool, toolHeader, toolHeaders) {
   if (trigger === 'mouseenter') {
 
-    // Hide panel when hover leaves tool header, unless hovering to panel
+    // Hide panel when hover leaves tool header, if new target element
+    // is part of the tools UI (and not the panel itself)
     toolHeader.addEventListener('mouseleave', event => {
       const toElement = event.toElement;
       const toId = toElement.id;
       const panelElement = document.querySelector('.ideo-tool-panel');
+      const toolsElement = document.querySelector('#tools');
       if (
+        toolsElement.contains(toElement) &&
         panelElement && !panelElement.contains(toElement) &&
         toId !== tool
       ) {
-        console.log('deactivate from handleHideForHoverables')
         deactivate(toolHeaders);
         panelElement.remove();
       }
-
-      console.log('')
-      console.log('')
-      console.log('')
     });
-
-    // if ()
   }
 }
 
@@ -312,9 +281,12 @@ function handleToolClick(ideo) {
       const tool = toolHeader.id.split('-')[0];
       const panel = getPanel(tool, ideo);
 
-      document.querySelector('#gear').insertAdjacentHTML('beforeend', panel);
-
-      handleHideForHoverables(trigger, tool, toolHeader, toolHeaders);
+      if (trigger === 'mouseenter') {
+        toolHeader.insertAdjacentHTML('beforeend', panel);
+        handleHideForHoverables(trigger, tool, toolHeader, toolHeaders);
+      } else {
+        document.querySelector('#gear').insertAdjacentHTML('beforeend', panel);
+      }
     });
   });
